Fixed assisted threading params to not use uint8_t as this causes issue with the python data struct types;

+++ Core/Inc/Ramps.h
@@ -89,11 +89,11 @@ typedef struct {
 
   // Assisted Threading support:
   uint32_t threadDesiredSteps; // host writes absolute desired steps
-  uint8_t threadRequest;      // host writes 1 to request latch+wait
-  uint8_t threadReset;         // host writes 1 to reset the threading state
+  uint16_t threadRequest;      // host writes 1 to request latch+wait
+  uint16_t threadReset;         // host writes 1 to reset the threading state
   uint16_t threadSpindleIndex; // host sets which scale to watch
   uint32_t threadPhaseRef;     // firmware stores latched phase
-  uint8_t threadHasPhase;      // 0 = not latched yet, 1 = latched
+  uint16_t threadHasPhase;      // 0 = not latched yet, 1 = latched
   uint16_t threadEnabled;      // firmware uses as waiting flag
   uint16_t threadTolerance;    // encoder-count tolerance for phase matching
 } fastData_t;
